third_party/blink/web_tests/external/wpt/css/css-fonts/test_font_feature_values_parsing-expected.txt

This is a testharness.js-based test.
Found 7 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] no value definitions in serialization - @font-feature-values bongo { @blah @styleset { abc: 1 2 3; } }
  assert_equals: cssText serialization should have no value defintions -  expected "" but got "@styleset { abc: 1 2 3; }"
[FAIL] basic parse tests - @font-feature-values bongo { @blah } @styleset { abc: 1 2 3; } }
  assert_true: unexpected syntax error expected true got false
[FAIL] no value definitions in serialization - @font-feature-values bongo { @blah , @styleset { abc: 1 2 3; } }
  assert_equals: cssText serialization should have no value defintions -  expected "" but got "@styleset { abc: 1 2 3; }"
[FAIL] serialization check - @font-feature-values bongo { \n@styleset\n  { blah: 3; super-blah: 4 5;\n  more-blah: 5 6 7;\n } }
  assert_equals: canonical cssText serialization doesn't match expected "@styleset { blah: 3; super-blah: 4 5; more-blah: 5 6 7; }" but got "@styleset { blah: 3; more-blah: 5 6 7; super-blah: 4 5; }"
[FAIL] serialization check - @font-feature-values bongo { \n@styleset\n  {\n blah:\n 3\n;\n super-blah:\n 4\n 5\n;\n  more-blah:\n 5 6\n 7;\n } }
  assert_equals: canonical cssText serialization doesn't match expected "@styleset { blah: 3; super-blah: 4 5; more-blah: 5 6 7; }" but got "@styleset { blah: 3; more-blah: 5 6 7; super-blah: 4 5; }"
[FAIL] serialization check - @font-feature-values bongo { @styleset { blah: 1; blah: 1; } }
  assert_equals: canonical cssText serialization doesn't match expected "@styleset { blah: 1; blah: 1; }" but got "@styleset { blah: 1; }"
[FAIL] serialization check - @font-feature-values bongo { @styleset { blah: 1; de-blah: 1; blah: 2; } }
  assert_equals: canonical cssText serialization doesn't match expected "@styleset { blah: 1; de-blah: 1; blah: 2; }" but got "@styleset { blah: 2; de-blah: 1; }"
Harness: the test ran to completion.
